requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
celery==5.3.4
kombu==5.3.4
billiard==4.2.0
amqp==5.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
